- add tests and retain local copy of element
