+++ src/Controls/src/Core/Compatibility/Handlers/Android/FrameRenderer.cs
@@ -46,6 +46,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 		GradientDrawable? _backgroundDrawable;
 		private IMauiContext? _mauiContext;
 		ViewHandlerDelegator<Frame> _viewHandlerWrapper;
+		VisualElement _element;
 		public event EventHandler<VisualElementChangedEventArgs>? ElementChanged;
 		public event EventHandler<PropertyChangedEventArgs>? ElementPropertyChanged;
 
@@ -63,6 +64,8 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 			{
 				if (value != null)
 					(this as IPlatformViewHandler).SetVirtualView(value);
+
+				_element = value;
 			}
 		}
 
@@ -337,8 +340,11 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 		void IElementHandler.SetMauiContext(IMauiContext mauiContext) =>
 			_mauiContext = mauiContext;
 
-		void IElementHandler.SetVirtualView(Maui.IElement view) =>
+		void IElementHandler.SetVirtualView(Maui.IElement view)
+		{
 			_viewHandlerWrapper.SetVirtualView(view, OnElementChanged, false);
+			_element = view as VisualElement;
+		}
 
 		void IElementHandler.UpdateValue(string property)
 		{
