+++ src/Controls/src/Core/Compatibility/Handlers/TabbedPage/iOS/TabbedRenderer.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 using Microsoft.Maui.Controls.Internals;
 using Microsoft.Maui.Controls.Platform;
 using Microsoft.Maui.Graphics;
@@ -25,6 +26,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 		bool? _defaultBarTranslucent;
 		IMauiContext _mauiContext;
 		UITabBarAppearance _tabBarAppearance;
+		VisualElement _element;
 
 		IMauiContext MauiContext => _mauiContext;
 		public static IPropertyMapper<TabbedPage, TabbedRenderer> Mapper = new PropertyMapper<TabbedPage, TabbedRenderer>(TabbedViewHandler.ViewMapper);
@@ -54,7 +56,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 			get { return (TabbedPage)Element; }
 		}
 
-		public VisualElement Element => _viewHandlerWrapper.Element;
+		public VisualElement Element => _viewHandlerWrapper.Element ?? _element;
 
 		public event EventHandler<VisualElementChangedEventArgs> ElementChanged;
 
@@ -69,6 +71,7 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 		public void SetElement(VisualElement element)
 		{
 			_viewHandlerWrapper.SetVirtualView(element, OnElementChanged, false);
+			_element = element;
 
 			FinishedCustomizingViewControllers += HandleFinishedCustomizingViewControllers;
 			Tabbed.PropertyChanged += OnPropertyChanged;
@@ -128,13 +131,10 @@ namespace Microsoft.Maui.Controls.Handlers.Compatibility
 				_tabBarAppearance = null;
 
 				Page?.SendDisappearing();
-				
-				if (Tabbed != null)
-				{
-					Tabbed.PropertyChanged -= OnPropertyChanged;
-					Tabbed.PagesChanged -= OnPagesChanged;
-				}
-				
+
+				Tabbed.PropertyChanged -= OnPropertyChanged;
+				Tabbed.PagesChanged -= OnPagesChanged;
+
 				FinishedCustomizingViewControllers -= HandleFinishedCustomizingViewControllers;
 			}
